Moved models to own package
