+++ service-admin/src/main/java/org/oskari/admin/model/MapLayerAdminOutput.java
@@ -1,4 +1,4 @@
-package org.oskari.admin;
+package org.oskari.admin.model;
 
 import java.util.Map;
