+++ service-admin/src/test/resources/org/oskari/admin/model/MapLayer-expected.json
